+++ src/components/EditTask/EditTask.tsx
@@ -29,6 +29,7 @@ type State = {
   taskComments: Array<TaskCommentType>;
   isEditAnyFields: boolean;
   boardId: string;
+  isClickOnBoardsPick: boolean;
 };
 
 class EditTask extends React.Component<Props, State> {
@@ -45,6 +46,7 @@ class EditTask extends React.Component<Props, State> {
       isTypeComment: false,
       isEditAnyFields: false,
       boardId: props.boardId,
+      isClickOnBoardsPick: false,
     };
   }
 
@@ -80,12 +82,11 @@ class EditTask extends React.Component<Props, State> {
     if (isCancel) {
       this.setState({ isEditTaskDescription: false });
     } else {
-      this.setState(
-        {
-          isEditTaskDescription: false,
-          taskDescription: value,
-          isEditAnyFields: true
-        });
+      this.setState({
+        isEditTaskDescription: false,
+        taskDescription: value,
+        isEditAnyFields: true,
+      });
     }
   };
 
@@ -100,25 +101,28 @@ class EditTask extends React.Component<Props, State> {
       };
       const prevTaskComments = this.state.taskComments;
       const newTaskComments = [...prevTaskComments, newComment];
-      this.setState(
-        {
-          isTypeComment: false,
-          taskComments: newTaskComments,
-          isEditAnyFields: true
-        });
+      this.setState({
+        isTypeComment: false,
+        taskComments: newTaskComments,
+        isEditAnyFields: true,
+      });
     }
   };
 
-  startEditTaskText = (e:React.MouseEvent<HTMLParagraphElement>) => {
+  startEditTaskText = (e: React.MouseEvent<HTMLParagraphElement>) => {
     this.setState({ isEditTaskText: true });
   };
 
-  startEditTaskDescription = (e:React.MouseEvent<HTMLParagraphElement>) => {
+  startEditTaskDescription = (e: React.MouseEvent<HTMLParagraphElement>) => {
     this.setState({ isEditTaskDescription: true, isTypeComment: false });
   };
 
-  startTypeComment = (e:React.MouseEvent<HTMLParagraphElement>) => {
+  startTypeComment = (e: React.MouseEvent<HTMLParagraphElement>) => {
     this.setState({ isTypeComment: true, isEditTaskDescription: false });
+  };
+
+  startChangingBoard = (e: React.MouseEvent<HTMLDivElement>) => {
+    this.setState({ isClickOnBoardsPick: true });
   };
 
   collectTaskData = (): EditTaskDataType => {
@@ -149,7 +153,7 @@ class EditTask extends React.Component<Props, State> {
   changeCurrentBoardId = (e: React.MouseEvent<HTMLDivElement>, newId: string) => {
     e.preventDefault();
     e.stopPropagation();
-    this.setState({ boardId: newId, isEditAnyFields: true });
+    this.setState({ boardId: newId, isEditAnyFields: true, isClickOnBoardsPick: false });
   };
 
   render() {
@@ -162,41 +166,50 @@ class EditTask extends React.Component<Props, State> {
       updateTime,
       taskComments,
       isTypeComment,
-      boardId } = this.state;
+      boardId,
+      isClickOnBoardsPick,
+    } = this.state;
 
     const { boardsInfo } = this.props;
 
-    const viewTaskText = isEditTaskText
-      ? <AddTextValue
-        startValue={taskText}
-        returnValueAction={this.getNewTaskText}
-      />
-      : <p className='EditTask-taskText' onClick={this.startEditTaskText}>{taskText}</p>;
+    const viewTaskText = isEditTaskText ? (
+      <AddTextValue startValue={taskText} returnValueAction={this.getNewTaskText} />
+    ) : (
+      <p className="EditTask-taskText" onClick={this.startEditTaskText}>
+        {taskText}
+      </p>
+    );
 
-    const viewTaskDescription = isEditTaskDescription
-      ? <AddFormatText
+    const viewTaskDescription = isEditTaskDescription ? (
+      <AddFormatText
         startValue={taskDescription}
         returnValueAction={this.getNewTaskDescription}
-        placeholder='You can add a task description here...'
+        placeholder="You can add a task description here..."
       />
-      : <p className='EditTask-description' onClick={this.startEditTaskDescription}>
+    ) : (
+      <p className="EditTask-description" onClick={this.startEditTaskDescription}>
         {taskDescription === '' ? 'Click to add a description...' : taskDescription}
-      </p>;
+      </p>
+    );
 
-    const viewTaskComments = taskComments.map(task => {
+    const viewTaskComments = taskComments.map((task) => {
       return <TaskComment {...task} key={task.commentDate} />;
     });
 
-    const viewNewComment = isTypeComment
-      ? <AddFormatText
-        startValue=''
+    const viewNewComment = isTypeComment ? (
+      <AddFormatText
+        startValue=""
         returnValueAction={this.getNewCommentText}
-        placeholder='Type your comment to the task'
+        placeholder="Type your comment to the task"
       />
-      : <p className='EditTask-newComment' onClick={this.startTypeComment}>Add a comment...</p>;
+    ) : (
+      <p className="EditTask-newComment" onClick={this.startTypeComment}>
+        Add a comment...
+      </p>
+    );
 
     let currentBoardName = '';
-    const otherBoards = boardsInfo.filter(item => {
+    const otherBoards = boardsInfo.filter((item) => {
       if (item.boardId === boardId) {
         currentBoardName = item.boardName;
         return false;
@@ -204,43 +217,53 @@ class EditTask extends React.Component<Props, State> {
       return true;
     });
 
-    const viewBoard = <div className='EditTask-board'>
-      {currentBoardName}  <span className='rotate'>{'>'}</span>
-      <div className='options'>{otherBoards.map(item => {
-        return <div 
-          key={item.boardName}
-          onClick={(e) => this.changeCurrentBoardId(e, item.boardId)}
-        >
+    const alternativeBoards = otherBoards.map((item) => {
+      return (
+        <div key={item.boardName} onClick={(e) => this.changeCurrentBoardId(e, item.boardId)}>
           {item.boardName}
         </div>
-      })}
-      </div>
-    </div>
+      );
+    });
 
-    return(
-      <div className='EditTask'>
-        <div className='EditTask-header'>
-          <div onClick={this.hendleClickCloseButton} role='button' tabIndex={0}>
-            <img src={Cancel} alt='Close window' />
+    const viewAlternativeBoard = isClickOnBoardsPick ? (
+      <div className="EditTask-optionsContainer">
+        <div className="options">{alternativeBoards}</div>
+      </div>
+    ) : null;
+
+    const viewBoard = (
+      <div className="EditTask-board" onClick={this.startChangingBoard}>
+        {currentBoardName} <span className="rotate">{'>'}</span>
+        {viewAlternativeBoard}
+      </div>
+    );
+
+    return (
+      <div className="EditTask">
+        <div className="EditTask-header">
+          <div onClick={this.hendleClickCloseButton} role="button" tabIndex={0}>
+            <img src={Cancel} alt="Close window" />
           </div>
         </div>
-        <div className='EditTask-items'>
-          <div className='left'>
+        <div className="EditTask-items">
+          <div className="left">
             {viewTaskText}
-            <div className='EditTask-taskDescription'>
+            <div className="EditTask-taskDescription">
               <p>Description</p>
               {viewTaskDescription}
             </div>
-            <div className='EditTask-comments'>
+            <div className="EditTask-comments">
               <p>Activity</p>
-              <p>Show: <span>Comments</span></p>
+              <p>
+                Show: <span>Comments</span>
+              </p>
               {viewTaskComments}
               {viewNewComment}
             </div>
           </div>
-          <div className='right'>
+          <div className="right">
             {viewBoard}
-            <div className='EditTask-taskDate'>
+            <div className="EditTask-taskDate">
               <p>Created: {getCurrentDateAsString(createTime)}</p>
               <p>{updateTime ? `Updated: ${getCurrentDateAsString(updateTime)}` : ''}</p>
             </div>