+++ src/components/Warnings/DeleteBoard/DeleteBoard.tsx
@@ -3,8 +3,7 @@ import './DeleteBoard.scss';
 import { DeleteWarningdProps } from '../../../types/types';
 
 const DeleteBoard: React.FC<DeleteWarningdProps> = ({ confirmAction, cancelAction }) => {
-
-  useEffect(()=> {
+  useEffect(() => {
     function pressKey(e: KeyboardEvent) {
       if (e.key === 'Escape') cancelAction(e);
     }