+++ src/containers/Board/BoardItem/BoardItem.tsx
@@ -116,8 +116,7 @@ class BoardItem extends React.Component<IProps, State> {
     }
   };
 
-  cancelDeleteBoard = (e: React.MouseEvent<HTMLButtonElement>
-  | KeyboardEvent): void => {
+  cancelDeleteBoard = (e: React.MouseEvent<HTMLButtonElement> | KeyboardEvent): void => {
     e.preventDefault();
     e.stopPropagation();
     this.props.triggerPopUp(false, null);
@@ -140,7 +139,15 @@ class BoardItem extends React.Component<IProps, State> {
   };
 
   render() {
-    const { boardName, boardHeight, updateBoards, id, tasks, triggerPopUp, boardsInfo } = this.props;
+    const {
+      boardName,
+      boardHeight,
+      updateBoards,
+      id,
+      tasks,
+      triggerPopUp,
+      boardsInfo,
+    } = this.props;
     const {
       isAddingTask,
       borderRef,