+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2020-05-30
+### Added
+- add edit task logic;
+- add the ability to comment task
+- add the ability to type task description
+- each task display create time and last update time (if able)
+- add the ability to change task's board in edit task mode
+
 ## [0.1.0] - 2020-05-06
 ### Added
 - Skeleton of the project;